requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
coverage==6.2
cryptography==36.0.0
Django==3.2.9
h11==0.12.0
idna==3.3
outcome==1.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
